lab3.1

+++ LibraryCPP/Tests/vector.cpp
@@ -13,7 +13,7 @@ int main()
     }
 
     for (size_t i = 0 ; i < vector_size(vector) ; ++i)
-        vector_set(vector, i, i);
+        vector_set(vector, i, static_cast<int>(i));
 
     for (size_t i = 0 ; i < vector_size(vector) ; ++i)
     {
